+++ packages/expo-file-system/src/ExponentFileSystemShim.ts
@@ -10,9 +10,6 @@ const platformModule: ExponentFileSystemModule = {
   get cacheDirectory(): string | null {
     return null;
   },
-  get bundledAssets(): string | null {
-    return null;
-  },
   get bundleDirectory(): string | null {
     return null;
   },